backend/requirements.txt

chardet==5.1.0
faiss-cpu==1.7.4
Flask-Cors==4.0.0
flask-session==0.5.0
gunicorn==21.2.0
langchain==0.0.272
openai==0.27.8
openapi-schema-pydantic==1.2.4
pip==21.2.3
pipdeptree==2.13.1
pymongo==4.6.0
python-dotenv==1.0.0
setuptools==57.4.0
tiktoken==0.5.1

